+++ django/applications/catmaid/static/js/widgets/connector-table.js
@@ -93,6 +93,8 @@
           {title: 'Incoming connectors', value: "postsynaptic_to"},
           {title: 'Outgoing connectors', value: "presynaptic_to"},
           {title: 'Gap junction connectors', value: "gapjunction_with"},
+          {title: 'Tight junction connectors', value: "tightjunction_with"},
+          {title: 'Desmosome connectors', value: "desmosome_with"},
           {title: 'Abutting connectors', value: "abutting"},
           {title: 'Attachment connectors', value: "attached_to"}],
           "presynaptic_to");